requirements.txt
# Core multi-agent and LLM
agno>=2.2.12
openai>=1.0.0

# Web/API
fastapi>=0.115.0
uvicorn[standard]>=0.23.0

# Observability queries
prometheus-api-client>=0.5.4
requests>=2.31.0

# gRPC and protobuf
grpcio>=1.60.0
grpcio-tools>=1.60.0
grpcio-reflection>=1.60.0
protobuf>=4.25.0

# Database
psycopg2-binary>=2.9.9

# Docker SDK
docker>=6.1.3

# ORM para AgentOS (sqlite)
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Config & utils
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.3

# Metrics
prometheus-client>=0.19.0

# OpenTelemetry
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
opentelemetry-exporter-otlp-proto-http>=1.21.0
opentelemetry-instrumentation-grpc>=0.42b0
opentelemetry-instrumentation-asyncio>=0.42b0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-logging>=0.42b0


